pyproject.toml
[tool.poetry]
name = "yolo-ai-cmdbot"
version = "0.1.0"
description = ""
authors = ["mohamed-aziz <me@aziz.tn>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.scripts]
yolo = 'src.yolo:main'

[tool.poetry.dependencies]
python = "^3.10"
termcolor = "2.2.0"
colorama = "0.4.4"
python-dotenv = "1.0.0"
distro = "1.7.0"
pyyaml = "6.0.1"
pyperclip = "1.8.2"
litellm = "^1.14.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
